// src/app/core/api-client/models/comment-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CommentResponseDto {
  authorId: string;
  authorName?: string;
  content: string;
  createdAt: string;
  entityId: string;
  entityType: string;
  id: string;
  parentId?: string;
  replies: Array<CommentResponseDto>;
  updatedAt: string;
}