// src/app/core/api-client/models/create-comment-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateCommentDto {

  /**
   * The content of the comment (emojis supported)
   */
  content: string;

  /**
   * The ID of the entity being commented on
   */
  entityId: string;

  /**
   * The type of entity being commented on
   */
  entityType: string;

  /**
   * The ID of the parent comment if this is a reply
   */
  parentId?: string;
}